move onRecordAvailable to RealtimeConnection
