+++ realtime-core/utils.ts
@@ -220,3 +220,12 @@ export const stringify = (obj: object): string => {
     return value;
   });
 };
+
+export const blobToBase64 = (blob: Blob): Promise<string | null> => {
+  return new Promise((resolve) => {
+    const reader = new FileReader();
+    reader.onloadend = () =>
+      resolve(reader.result?.toString().split(",")[1] || null);
+    reader.readAsDataURL(blob);
+  });
+};
