--- realtime-react/hooks/utils.ts
@@ -1,8 +1,0 @@
-export const blobToBase64 = (blob: Blob): Promise<string | null> => {
-  return new Promise((resolve) => {
-    const reader = new FileReader();
-    reader.onloadend = () =>
-      resolve(reader.result?.toString().split(",")[1] || null);
-    reader.readAsDataURL(blob);
-  });
-};
